% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databycountry.R
\docType{class}
\name{databycountry-class}
\alias{databycountry-class}
\alias{databycountry}
\title{INQStats data query}
\description{
Returns the data from INQStats.
Further information about  INQStats, e.g. how to get api key, what data is available, can be found on \url{http://blog.inqubu.com/inqstats-open-api-published-to-get-demographic-data}
}
\section{Fields}{

\describe{
\item{\code{api_key}}{a key received from INQStats.}

\item{\code{country_code}}{a string of ISO 3166-1 country codes, e.g. "us,gb,se".}

\item{\code{data}}{a data, presented as a vector, e.g. c("population","birth_rate").}

\item{\code{years}}{a year range, as a vector, e.g.  1990:2002.}

\item{\code{result}}{a data frame.}
}}


\examples{
databycountry$new(api_key= "97c1f72e01dd5ba3", country_code = "us,lt,se", data = c("population","birth_rate","co2_emissions","fifa","hdi"),years = 1990:2015)$result
}
