\name{INQStatsR-package}
\alias{INQStatsR-package}
\alias{INQStatsR}
\docType{package}
\title{
\packageTitle{INQStatsR}
}
\description{
\packageDescription{INQStatsR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{INQStatsR}
\packageIndices{INQStatsR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{INQStatsR}

Maintainer: \packageMaintainer{INQStatsR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
