% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runShinyapp}
\alias{runShinyapp}
\title{INQStats Shiny visualization}
\usage{
runShinyapp(data)
}
\description{
Visualize a time series of the data extracted from INQStats API
}
\section{Fields}{

\describe{
\item{\code{results}}{from a data frame.}
}}

