\name{INQStatsR-package}
\alias{INQStatsR-package}
\docType{package}
\title{
\packageTitle{INQStatsR}
}
\description{
\packageDescription{INQStatsR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{INQStatsR}
\packageIndices{INQStatsR}

}
\author{
\packageAuthor{INQStatsR}

Maintainer: \packageMaintainer{INQStatsR}
}
\references{
\url{http://blog.inqubu.com/inqstats-open-api-published-to-get-demographic-data}
}

\keyword{ package }
\seealso{
\url{https://github.com/henkar91/AdvRprogr_lab5}
}
\examples{

}
